code cleanup and proofread of docs
