+++ docs/src/dev-docs/utils/omegaconf.rst
@@ -1,0 +1,10 @@
+Custom omegaconf functions
+==========================
+
+Resolvers to handle special fields in our configs as well as the expansion/completion of
+the dataset section.
+
+.. automodule:: metatensor.models.utils.omegaconf
+    :members:
+    :undoc-members:
+    :show-inheritance:
