+++ docs/src/getting-started/usage.rst
@@ -66,3 +66,7 @@ The sub-command to evaluate a already trained model is
 .. literalinclude:: ../../../examples/usage.sh
     :language: bash
     :lines: 9-
+
+
+In the next tutorials we show how adjust the dataset section of ``options.yaml`` file
+to use it for your own datasets.
