--- src/metatensor/models/cli/conf/dataset/gradients_avail.yaml
@@ -1,3 +1,0 @@
-forces: off
-stress: off
-virial: off
