--- src/metatensor/models/cli/conf/dataset/structures.yaml
@@ -1,3 +1,0 @@
-read_from:
-file_format:
-unit:
