--- src/metatensor/models/cli/conf/dataset/targets.yaml
@@ -1,5 +1,0 @@
-quantity: energy
-read_from: ${...structures.read_from}
-file_format:
-key:
-unit:
