+++ tests/cli/test_train_model.py
@@ -3,9 +3,6 @@ import subprocess
 from pathlib import Path
 
 import pytest
-from omegaconf import OmegaConf
-
-from metatensor.models.cli.train_model import expand_dataset_config
 
 
 RESOURCES_PATH = Path(__file__).parent.resolve() / ".." / "resources"
@@ -52,137 +49,3 @@ def test_hydra_arguments():
 
 
 # TODO: test split of train/test/validation using floats and combinations of these.
-
-
-def test_expand_dataset_config():
-    file_name = "foo.xyz"
-    file_format = ".xyz"
-
-    structure_section = {"read_from": file_name, "unit": "angstrom"}
-
-    target_section = {
-        "quantity": "energy",
-        "forces": file_name,
-        "virial": file_name,
-        "bar": {"read_from": "my_grad.dat", "key": "foo"},
-    }
-
-    conf = {
-        "structures": structure_section,
-        "targets": {"energy": target_section, "energy2": target_section},
-    }
-
-    conf_expanded = expand_dataset_config(OmegaConf.create(conf))
-
-    assert conf_expanded["structures"]["read_from"] == file_name
-    assert conf_expanded["structures"]["file_format"] == file_format
-    assert conf_expanded["structures"]["unit"] == "angstrom"
-
-    targets_conf = conf_expanded["targets"]
-    assert len(targets_conf) == 2
-
-    assert targets_conf["energy"]["quantity"] == "energy"
-    assert targets_conf["energy"]["read_from"] == file_name
-    assert targets_conf["energy"]["file_format"] == file_format
-    assert targets_conf["energy"]["file_format"] == file_format
-    assert targets_conf["energy"]["key"] == "energy"
-    assert targets_conf["energy"]["unit"] is None
-
-    for gradient in ["forces", "virial"]:
-        assert targets_conf["energy"][gradient]["read_from"] == file_name
-        assert targets_conf["energy"][gradient]["file_format"] == file_format
-        assert targets_conf["energy"][gradient]["key"] == gradient
-
-    assert targets_conf["energy"]["bar"]["read_from"] == "my_grad.dat"
-    assert targets_conf["energy"]["bar"]["key"] == "foo"
-
-    # If a virial is parsed as in the conf above the by default enabled section "stress"
-    # should be disabled automatically
-    assert targets_conf["energy"]["stress"] is False
-
-    assert targets_conf["energy2"]["key"] == "energy2"
-    assert targets_conf["energy"]["quantity"] == "energy"
-
-
-def test_expand_dataset_config_not_energy():
-    file_name = "foo.xyz"
-
-    structure_section = {"read_from": file_name, "unit": "angstrom"}
-
-    target_section = {
-        "quantity": "my_dipole_moment",
-    }
-
-    conf = {
-        "structures": structure_section,
-        "targets": {"dipole_moment": target_section},
-    }
-
-    conf_expanded = expand_dataset_config(OmegaConf.create(conf))
-
-    assert conf_expanded["targets"]["dipole_moment"]["key"] == "dipole_moment"
-    assert conf_expanded["targets"]["dipole_moment"]["quantity"] == "my_dipole_moment"
-    assert conf_expanded["targets"]["dipole_moment"]["forces"] is False
-    assert conf_expanded["targets"]["dipole_moment"]["stress"] is False
-    assert conf_expanded["targets"]["dipole_moment"]["virial"] is False
-
-
-def test_expand_dataset_config_min():
-    file_name = "dataset.dat"
-    file_format = ".dat"
-
-    conf_expanded = expand_dataset_config(file_name)
-
-    assert conf_expanded["structures"]["read_from"] == file_name
-    assert conf_expanded["structures"]["file_format"] == file_format
-
-    targets_conf = conf_expanded["targets"]
-    assert targets_conf["energy"]["quantity"] == "energy"
-    assert targets_conf["energy"]["read_from"] == file_name
-    assert targets_conf["energy"]["file_format"] == file_format
-    assert targets_conf["energy"]["file_format"] == file_format
-    assert targets_conf["energy"]["key"] == "energy"
-    assert targets_conf["energy"]["unit"] is None
-
-    for gradient in ["forces", "stress"]:
-        assert targets_conf["energy"][gradient]["read_from"] == file_name
-        assert targets_conf["energy"][gradient]["file_format"] == file_format
-        assert targets_conf["energy"][gradient]["key"] == gradient
-
-    assert targets_conf["energy"]["virial"] is False
-
-
-def test_expand_dataset_config_error():
-    file_name = "foo.xyz"
-
-    conf = {
-        "structures": file_name,
-        "targets": {
-            "energy": {
-                "virial": file_name,
-                "stress": {"read_from": file_name, "key": "foo"},
-            }
-        },
-    }
-
-    with pytest.raises(
-        ValueError, match="Cannot perform training with respect to virials and stress"
-    ):
-        expand_dataset_config(OmegaConf.create(conf))
-
-
-def test_expand_dataset_gradient_true():
-    conf = {
-        "structures": "foo.xyz",
-        "targets": {
-            "energy": {
-                "virial": True,
-                "stress": False,
-            }
-        },
-    }
-
-    conf_expanded = expand_dataset_config(OmegaConf.create(conf))
-
-    assert conf_expanded["targets"]["energy"]["stress"] is False
-    conf_expanded["targets"]["energy"]["virial"]["read_from"]
